Fix Maven errors when accessing .qvto models
This is a quick fix for the missing .qvto models when trying to access .qvto models the same way as in eclipse after building this Repository with Maven. 
I am reffering to SPDAdjustorModule.java in the org.palladiosimulator.analyzer.slingshot.behavior.spd.adjustment module. 
Accessing MainTransformations.qvto with MAIN_QVTO_FILE = "platform:/plugin/org.palladiosimulator.spd.semantic.transformations/transformations/spd/MainTransformation.qvto" works fine when working with eclipse. But this model is not in the same location when trying the same with Maven. 

By adding these files in this src/main/resource folder, we ensure, that Maven puts the resources at the same location as Eclipse. Therefore the platform link above will work again.
These changes are temporary and best be changed in a different way than I did with this quick fix. My ideas would be to either provide something like a pathmap to this transformations module or to change the SPDAdjustorModule.java class, so that it will work for both Maven and Eclipse